requirements.txt
streamlit>=1.31.0
ultralytics>=8.1.25
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0